Add timeout to codesigning, also move it just before its needed
The keychain created in Import Certificates seems to lock after a period of time, so moving the installation of the certificates after the code build just before they are needed.  The 5 minute timeout was added due to a situation I was facing where a prompt to enter the keychain passwork had popped up and the step kept running until the whole workflow timed out.